requirements.txt
# pip install -U -r requirements.txt
Cython
matplotlib>=3.2.2
numpy==1.19.5
opencv-python
pillow
easydict
# pycocotools>=2.0
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
#torch>=1.6.0
#torchvision>=0.7.0
tqdm>=4.41.0
moviepy
shapely
seaborn
openvino==2021.4.0
scikit-image
filterpy
polygon_manager
python-dotenv
pandas==1.4.0
azure-storage-blob
mysql-connector-python==8.0.29
